Move user following suggestions to internal API
