+++ app/javascript/sidebar-widget/SidebarWidget.jsx
@@ -29,7 +29,7 @@ class SidebarWidget extends Component {
 
   getSuggestedUsers() {
     const { tagInfo } = this.state;
-    fetch(`/api/users?state=sidebar_suggestions&tag=${tagInfo.name}`, {
+    fetch(`/users?state=sidebar_suggestions&tag=${tagInfo.name}`, {
       headers: {
         Accept: 'application/json',
         'Content-Type': 'application/json',
